+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostBotLikesPlugin/listeners/ForceAttachLike.kt
@@ -1,6 +1,6 @@
 package com.github.insanusmokrassar.AutoPostBotLikesPlugin.listeners
 
-import com.github.insanusmokrassar.AutoPostBotLikesPlugin.database.LikesPluginRegisteredLikesMessagesTable
+import com.github.insanusmokrassar.AutoPostBotLikesPlugin.database.LikesPluginMessagesTable
 import com.github.insanusmokrassar.AutoPostBotLikesPlugin.utils.extensions.AdminsHolder
 import com.github.insanusmokrassar.AutoPostTelegramBot.allMessagesListener
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.subscribe
@@ -12,18 +12,18 @@ import com.github.insanusmokrassar.TelegramBotAPI.types.message.ForwardedFromCha
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.CommonMessage
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.FromUserMessage
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.TextContent
-import com.github.insanusmokrassar.TelegramBotAPI.utils.extensions.executeAsync
 import com.github.insanusmokrassar.TelegramBotAPI.utils.extensions.executeUnsafe
 import org.joda.time.DateTime
 import java.lang.ref.WeakReference
 
-private val commandRegex: Regex = Regex("^/attachTargetLike [\\d]+$")
-private const val commandTemplate: String = "/attachTargetLike %d"
+private val commandRegex: Regex = Regex("^/attachLikes$")
+private const val commandTemplate: String = "/attachLikes"
 
 internal fun enableDetectLikesAttachmentMessages(
     adminsHolder: AdminsHolder,
     targetChatId: ChatId,
-    likesPluginRegisteredLikesMessagesTable: LikesPluginRegisteredLikesMessagesTable,
+    likesPluginMessagesTable: LikesPluginMessagesTable,
+    likesGroupsRegistrator: LikesGroupsRegistrator,
     botWR: WeakReference<RequestsExecutor>
 ) {
     allMessagesListener.subscribe {
@@ -34,11 +34,11 @@ internal fun enableDetectLikesAttachmentMessages(
         when (forwarded) {
             is ForwardedFromChannelMessage -> {
                 val originalMessageId = forwarded.messageId
-                if (forwarded.channelChat.id == targetChatId && adminsHolder.contains(userId) && !likesPluginRegisteredLikesMessagesTable.contains(originalMessageId)) {
+                if (forwarded.channelChat.id == targetChatId && adminsHolder.contains(userId) && !likesPluginMessagesTable.contains(originalMessageId)) {
                     botWR.get() ?.executeUnsafe(
                         SendMessage(
                             message.chat.id,
-                            "Ok, send me `${commandTemplate.format(originalMessageId)}` for attach post likes",
+                            "Ok, reply this message and send me $commandTemplate",
                             MarkdownParseMode
                         )
                     )
@@ -46,30 +46,19 @@ internal fun enableDetectLikesAttachmentMessages(
             }
             else -> {
                 (message.content as? TextContent) ?.also {
-                    if (commandRegex.matches(it.text) && adminsHolder.contains(userId)) {
-                        val messageId = it.text.split(" ")[1].toLong()
+                    val reply = message.replyTo
+                    if (commandRegex.matches(it.text) && reply != null && adminsHolder.contains(userId)) {
 
-                        if (messageId !in likesPluginRegisteredLikesMessagesTable) {
-                            likesPluginRegisteredLikesMessagesTable.registerMessageId(
-                                messageId,
-                                DateTime.now()
-                            ).also {
-                                if (it) {
-                                    botWR.get() ?.executeUnsafe(
-                                        SendMessage(
-                                            message.chat.id,
-                                            "Likes was attached (can be showed with delay)"
-                                        )
-                                    )
-                                } else {
-                                    botWR.get() ?.executeUnsafe(
-                                        SendMessage(
-                                            message.chat.id,
-                                            "Likes was not attached (can be already attached)"
-                                        )
-                                    )
-                                }
-                            }
+                        if (reply.chat.id == targetChatId && reply.messageId !in likesPluginMessagesTable) {
+                            likesGroupsRegistrator.registerNewLikesGroup(
+                                listOf(reply)
+                            )
+                            botWR.get() ?.executeUnsafe(
+                                SendMessage(
+                                    message.chat.id,
+                                    "Likes was attached (can be showed with delay)"
+                                )
+                            )
                         }
                     }
                 }